# backend/docs/api-docs.json
{
  "openapi": "3.1.0",
  "info": { "title": "ticket-bus", "description": "", "version": "1.0.0" },
...
  "paths": {
...
    "/vehicles": {
      "get": {
        "summary": "Obtener todos los vehiculos y sus amenities",
        "deprecated": false,
        "description": "",
        "operationId": "getVehiclesAmenities",
        "tags": ["Vehicles", "Vehicles"],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Operación exitosa",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculos obtenidos con éxito!",
                      "body": {
                        "vehicles": [
                          {
                            "id": 1,
                            "number": 1,
                            "plate": "QRST90",
                            "totalSeats": 60,
                            "amenities": [
                              "Cargador USB",
                              "WiFi",
                              "Aire Acondicionado"
                            ]
                          },
                          {
                            "id": 2,
                            "number": 2,
                            "plate": "ABCD12",
                            "totalSeats": 40,
                            "amenities": ["WiFi", "TV"]
                          }
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Error interno del servidor",
            "content": {
              "application/json": {
                "schema": { "$ref": "#/components/schemas/ApiResponse" },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "error",
                      "code": 500,
                      "message": "Error al obtener los vehiculos!"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/vehicles/id/{id}": {
      "get": {
        "summary": "Obtener un vehiculos y sus amenities",
        "deprecated": false,
        "description": "",
        "operationId": "getVehicleAmenityById",
        "tags": ["Vehicles", "Vehicles"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID de Vehiculos",
            "required": true,
            "example": "1",
            "schema": { "type": "integer" }
          }
        ],
        "responses": {
          "200": {
            "description": "Operacion exitosa",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "code": { "type": "integer" },
                    "message": { "type": "string" },
                    "body": {
                      "type": "object",
                      "properties": {
                        "id": { "type": "integer" },
                        "number": { "type": "integer" },
                        "plate": { "type": "string" },
                        "totalSeats": { "type": "integer" },
                        "amenities": {
                          "type": "array",
                          "items": { "type": "string" }
                        }
                      },
                      "required": [
                        "id",
                        "number",
                        "plate",
                        "totalSeats",
                        "amenities"
                      ]
                    }
                  },
                  "required": ["status", "code", "message", "body"]
                },
                "examples": {
                  "1": {
                    "summary": "Success",
                    "value": {
                      "status": "success",
                      "code": 200,
                      "message": "Vehiculo obtenido con éxito!",
                      "body": {
                        "id": 1,
                        "number": 1,
                        "plate": "QRST90",
                        "totalSeats": 60,
                        "amenities": [
                          "Cargador USB",
                          "WiFi",
                          "Aire Acondicionado"
                        ]
                      }
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "Record Not Found",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "status": { "type": "string" },
                    "code": { "type": "integer" },
                    "message": { "type": "string" }
                  },
                  "required": ["status", "code", "message"]
                },
                "examples": {
                  "1": {
                    "summary": "Exception",
                    "value": {
                      "status": "fail",
                      "code": 404,
                      "message": "No se encontro ningún vehiculo"
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiResponse": {
        "type": "object",
        "properties": {
          "status": { "type": "string" },
          "code": { "type": "integer" },
          "message": { "type": "string" },
          "body": { "type": "object", "properties": {} }
        },
        "required": ["status", "code", "message"]
      }
    },
    "securitySchemes": {}
  },
  "servers": [
    { "url": "http://localhost:3000/api/v1", "description": "Develop Env" },
    {
      "url": "https://ticket-bus-api-dev.up.railway.app/api/v1",
      "description": "Prod Env"
    }
  ]
}